requirements.txt
amqp==5.3.1
aniso8601==10.0.1
annotated-types==0.7.0
async-timeout==5.0.1
bidict==0.23.1
billiard==4.2.1
blinker==1.9.0
celery==5.5.3
celery-types==0.23.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
dependency-injector==4.48.1
dnspython==2.7.0
eventlet==0.40.3
exceptiongroup==1.3.0
Flask==3.1.2
flask-cors==6.0.1
Flask-RESTful==0.3.10
Flask-SocketIO==5.5.1
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.5.4
MarkupSafe==3.0.2
packaging==25.0
pluggy==1.6.0
prompt_toolkit==3.0.52
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pytest==8.4.2
pytest-mock==3.15.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.2
python-socketio==5.13.0
pytz==2025.2
redis==6.4.0
requests==2.32.5
simple-websocket==1.1.0
six==1.17.0
tomli==2.2.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
watchdog
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
